#48 Disable duplicate line warnings from WNT messages and rename files
