+++ wrappers/python/wirepas_messaging/wnt/ws_api/authenticationmessages.py
@@ -1,3 +1,4 @@
+# pylint: disable=duplicate-code
 """
     Authentication
     ==============
