+++ wrappers/python/wirepas_messaging/wnt/ws_api/scratchpadstatusmessages.py
